+++ pkg/resource/logging-config/alloy-logging-config.go
@@ -82,6 +82,7 @@ func generateAlloyConfig(lc loggedcluster.Interface, observabilityBundleVersion 
 		BasicAuthUsernameEnvVarName string
 		BasicAuthPasswordEnvVarName string
 		SupportPodLogs              bool
+		InsecureSkipVerify          bool
 	}{
 		ClusterID:                   clusterName,
 		Installation:                lc.GetInstallationName(),
@@ -92,7 +93,8 @@ func generateAlloyConfig(lc loggedcluster.Interface, observabilityBundleVersion 
 		BasicAuthUsernameEnvVarName: loggingsecret.AlloyBasicAuthUsernameEnvVarName,
 		BasicAuthPasswordEnvVarName: loggingsecret.AlloyBasicAuthPasswordEnvVarName,
 		// Observability bundle in older versions do not support PodLogs
-		SupportPodLogs: observabilityBundleVersion.GE(supportPodLogs),
+		SupportPodLogs:     observabilityBundleVersion.GE(supportPodLogs),
+		InsecureSkipVerify: lc.IsInsecureCA(),
 	}
 
 	err := alloyLoggingTemplate.Execute(&values, data)